Fix: отдельние типов данных
