+++ src/types/offer-type/host.ts
@@ -1,0 +1,7 @@
+type Host = {
+  isPro : boolean;
+  name: string;
+  avatarUrl: string;
+};
+
+export type {Host};
